Moving the component as per PR comments
