+++ src/app/services/logger/logger.module.ts
@@ -2,7 +2,7 @@ import {NgModule, CUSTOM_ELEMENTS_SCHEMA} from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { environment } from 'src/environments/environment';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
-import { ExuiLoggerComponent } from './exui-logger/exui-logger.component';
+import { ExuiLoggerComponent } from '../../components/exui-logger/exui-logger.component';
 
 @NgModule({
     declarations: [
